requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2
pyjwt[crypto]
passlib[bcrypt]
email-validator
python-multipart